dist: Correctly namespace the python module
The python module we are trying to create is `bubbles`, which previously
existed at the root of the project and a sub-module of `backend`.

This commit moves the package into the correct namespace, which assists
with testing by ensuring the module is installed rather than just in
the system path.

We also keep tests out of the source directory to allow them to import
the installed module.

See: https://blog.ionelmc.ro/2014/05/25/python-packaging

Signed-off-by: Joshua Hesketh <josh@nitrotech.org>