deploy: Fix where the built bubbles frontend is expected
The build-container.sh script does the right thing by placing everything
into the dist folder that would be needed by the image. However, the rest
of the developer instructions continue by using `--shared_ceph_folder` to
help with development, which therefore expects the frontend to be built
in place.

This change leaves a copy of the built frontend so either method works.

Signed-off-by: Joshua Hesketh <jhesketh@suse.com>