Build bubbles image on non SUSE distributions
There is a bug in openSUSE images which prevents zypper to run in non
priviledge containers, however if you use a SUSE distribution you can
run zypper with out an error on unpriviledged containers. To build
bubbles buildah is used, however the build commands for docker or podman
don't allow priviledged building, therefore the build script has to be
run inside a vagrant box based on a SUSE distribution in oder to run the
build successfully.

The bug is found here:
https://bugzilla.opensuse.org/show_bug.cgi?id=1190670

Signed-off-by: Stephan Müller <smueller@suse.com>